+++ crates/nostr-relay-pool/src/relay/status.rs
@@ -5,22 +5,61 @@
 //! Relay status
 
 use core::fmt;
+use core::sync::atomic::{AtomicU8, Ordering};
+
+#[derive(Debug)]
+pub(super) struct AtomicRelayStatus {
+    value: AtomicU8,
+}
+
+impl Default for AtomicRelayStatus {
+    fn default() -> Self {
+        Self::new(RelayStatus::Initialized)
+    }
+}
+
+impl AtomicRelayStatus {
+    #[inline]
+    pub(super) fn new(status: RelayStatus) -> Self {
+        Self {
+            value: AtomicU8::new(status as u8),
+        }
+    }
+
+    #[inline]
+    pub fn set(&self, status: RelayStatus) {
+        self.value.store(status as u8, Ordering::SeqCst);
+    }
+
+    pub(super) fn load(&self) -> RelayStatus {
+        let val: u8 = self.value.load(Ordering::SeqCst);
+        match val {
+            0 => RelayStatus::Initialized,
+            1 => RelayStatus::Pending,
+            2 => RelayStatus::Connecting,
+            3 => RelayStatus::Connected,
+            4 => RelayStatus::Disconnected,
+            5 => RelayStatus::Terminated,
+            _ => unreachable!(),
+        }
+    }
+}
 
 /// Relay connection status
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub enum RelayStatus {
     /// Relay initialized
-    Initialized,
+    Initialized = 0,
     /// Pending
-    Pending,
+    Pending = 1,
     /// Connecting
-    Connecting,
+    Connecting = 2,
     /// Relay connected
-    Connected,
+    Connected = 3,
     /// Relay disconnected, will retry to connect again
-    Disconnected,
+    Disconnected = 4,
     /// Relay completely disconnected
-    Terminated,
+    Terminated = 5,
 }
 
 impl fmt::Display for RelayStatus {